move delete account UI to its own screen
