+++ home/screens/DeleteAccountScreen/DeleteAccountPermissionsQuery.query.graphql
