+++ home/screens/DeleteAccountScreen/OTPStep.tsx
@@ -9,14 +9,14 @@ import {
 } from 'expo-dev-client-components';
 import React, { Fragment, useEffect, useState } from 'react';
 
-import { Button } from '../../../components/Button';
+import { Button } from '../../components/Button';
 import {
   SecondFactorMethod,
   UserSecondFactorDevice,
   useSecondFactorDevicesQuery,
   useSendSmsotpToSecondFactorDeviceMutation,
-} from '../../../graphql/types';
-import { notEmpty } from '../../../utils/notEmpty';
+} from '../../graphql/types';
+import { notEmpty } from '../../utils/notEmpty';
 import { SMSDevice } from './SMSDevice';
 
 export type PartialUserSecondFactorDevice = Partial<
