+++ home/screens/DeleteAccountScreen/SecondFactorDevicesQuery.query.graphql
