+++ home/screens/DeleteAccountScreen/SendSMSOTPToSecondFactorDevice.mutation.graphql
