+++ home/screens/DeleteAccountScreen/UserPermissionData.fragment.graphql
